HFP-1907 Hide copyright button
For the video, audio and image content type, the copyright button
will be replaced by the changes to the metadata system throughout
HFP-1896 to HFP-1907. These changes allow to react to the boolean
property

  hideCopyright

in semantics. In consequence, not setting this property or setting
it to false will still show the copyright button which can make
sende in larger content types, while it can be hidden for just
wrapper like content types like video, audio and image.