HFP-1905 Remove Drupal's title field
The new language property 'untitled' will be used by changes in
the Drupal plugin to set a default title if none is available.
The Drupal plugin should also be updated along this commit because
the title field won't be visible for changes anymore, will always be
empty and would prevent saving content.