Entitle for metadata based on blocklist
The previous passlist was replaced by a blocklist, so it is now
ensured that old content types that don't require changes and were not
updated can also have the metadata button.

This may change, e.g. make a clear distinction between pre-/post-metadata
by bumping all content types and putting all of them in the blocklist