JI-2406 Fix CKEditor field collapsing
The CKEditor input field would sometimes collapse because it uses an
internal element to determine the height of the input field, and this
element was hidden and removed from the DOM so the height returned
was invalid, causing the input height to collapse to 0. Setting the
getBody element of CKEditor to an element that is always visible
and not removed fixes this measurement problem.