Guard against undefined data in translations where semantics has entries
It was found that in version 1.24.2 of "Interactive Video" there are
entries in the semantics.json file, where the corresponding entries in
the translations are missing. This breaks hard in h5peditor.js, function
updateCommonFieldsDefault, where the fields from the semantics entries
are iterated and `field` member of the translations entry is accessed
without checking, that is actually exists.