VA-426 Fix manual updating of CKE source element
This prevents multiple DOM updates and inconsistent behvaior
depending on the order of validate() and removeWysiwyg() calls.
Also, prevents seeing tags in .$input that is not allowed by the
semantics.
Refines the usage of table-overflow-protection and avoids it being
inserted if <mtable> tags are present, or simply the 'table' word.